Move API tasks out of transport implementation
API verison handling and error checking is not related to transport and
should not have to be implemented in all transports.
Instead make the client handle these things.

For versioning, this is handled through a context value.

Additionally add support for setting the max response body size to read
in a context value as well.