testutil: do not log response until close
This was breaking API's where the response headers are sent immediately
but the response body wasn't sent until later (such as in the wait API).

It's possible we could just wrap `t.Log` such that any bytes read from
the response would be logged as they are read, but this works for now.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>