logs: Make API more like exec API
This moves the logic in how to consume the logs into the client by
requiring the caller to pass in stdout and stderr streams directly and
automatically starts the goroutines to do the copying.

This also makes various test improvements which were pretty flimsy and
dependent on the right timings with the daemon.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>