Fix go1.20.6 host header for unix sockets
go1.20.6 made a security-related patch to header parsing that makes a
path in the host header invalid.
This breaks unix sockets (and probably named pipes on Windows).

This change sets the `req.Host` field to override the host header field
to a value go is happy with.
Namely it just replaces all instances of `/` with `_`

Signed-off-by: Brian Goff <cpuguy83@gmail.com>